This patch attempts to tweak OpenCL kernel for the follow aspects:
1. Reduce unnecessary memory access;
2. Remove non-reacheable code;
3. Specialize character-wise set;
4. Add loop unrolling hints;
5. Assume the messages no exceeding 17 exabytes and apply optimizations;

It is known to bring about 15% speedup on NVIDIA TITAN Xp.