@jserv 2020-07-31
Test environment:
*  Intel Xeon E5-2650 v4 @ 2.20GHz
* NVIDIA TITAN Xp

original OpenCL kernel:
```
go test -tags cl
ok  	github.com/inkeliz/nanopow	1.872s
```

new OpenCL kernel:
```
go test -tags cl
ok  	github.com/inkeliz/nanopow	1.515s
```