@jserv 2020-08-01
Thank @Inkeliz for performance matrix. Can you share benchmark tools which can generate reports? Meanwhile, I am working on https://github.com/nanocurrency/nano-work-server/pull/18 with identical improvements. However, I lack of reproducible tools.